+++ GraphicalUI/AppShell.xaml.cs
@@ -1,9 +1,0 @@
-﻿namespace GraphicalUI;
-
-public partial class AppShell : Shell
-{
-    public AppShell()
-    {
-        InitializeComponent();
-    }
-}