+++ GraphicalUI/Platforms/iOS/AppDelegate.cs
@@ -1,8 +1,0 @@
-﻿using Foundation;
-
-namespace GraphicalUI;
-[Register("AppDelegate")]
-public class AppDelegate : MauiUIApplicationDelegate
-{
-    protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
-}